+++ docs/changes.rst
@@ -25,6 +25,7 @@ Not yet released.
 * Added option for :ref:`enforcing-checks`.
 * Increased default validity of confirmation links.
 * Improved Matomo integration.
+* Fixed :ref:`check-translated` to correctly handle source string change.
 
 Weblate 3.10.3
 --------------